Add snapcraft.yaml defining the snap package and its components
The DCD package uses `classic` confinement to ensure that it can access
the resources of the host system it is installed on, e.g. any installed
D compilers, headers for auto-complete, and so on.

`Build-snaps' is used to ask snapcraft to install the DMD package 
in order to have a Dlang compiler to compile DCD, DMD is only 
used during the process of compiling the snap and is not included 
with this package

`DMD = /snap/bin/ dmd` is used to specify the location of the dmd 
executable contained in the DMD package, this is useful to avoid any 
problems caused by incorrect configuration of the PATH environment 
variable within the build environment.

`artifacts:` is used to specify the files to be copied to the installation folder 
of the snap, this we need to distribute only the useful executables 
(ie dcd-client and dcd-server).